+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Breaking: Remove phantomJs support only supports environments which have require
 * Potential Breaking: Remove Redundant top level mocha.suite
 * Drop the duration on messages if mocha returns undefined/null (for example skipped test) TeamCity will then use received timestamps to calculate duration
+* Support Show diff between expected and actual values
+
 
 TODO
 comparisonFailure service message attributes